Use the maintenance mode for Friendica
This is done to avoid errors with the database and be able to properly back it up. Else it will fail if you Friendica is an active instance and in the time that it takes to create a backup new entries will be added to the database. I tested it it works.